+++ CHANGELOG.md
@@ -8,6 +8,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
    * enabled document download in contract and consent
    * Delete document for step 1 , step 2 and step 3
    * Changed PDF uploads from mandatory to optional and changed dropzone parameter size
+   * Added privacy policy to app page
 * Service Release Process
    * Create Service, fetch service and Update service apis added
 * AppDetails